fixed service since paginate isn't duplicate
